pyproject.toml
[project]
name = "pybaml"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "autogluon-tabular[all]==1.3.1",
    "datasphere>=0.10.0",
    "docutils==0.21.2",
    "fastai>=2.8.4",
    "flaml==2.3.5",
    "imbalanced-learn==0.12.3",
    "matplotlib==3.7.5",
    "numpy==1.26.0",
    "openml==0.14.2",
    "pandas==2.2.3",
    "pytest==8.3.5",
    "seaborn==0.13.2",
    "setuptools>=80.9.0",
    "torch>=2.2,<2.7",
    "tqdm==4.66.5",
    "h2o>=3.46.0.9",
    "loguru>=0.7.3",
    "polars==1.36.1",
    "pyarrow==22.0.0",
    "install-jdk==1.1.0",
    "ray>=2.10.0,<2.45.0"
]

[global]
extra-index-url = "https://download.pytorch.org/whl/cpu"
